revert 79a24ec just for other-stuff/gittalk.notes
creole2md automatic changes in 79a24ec are ok for everything else but not for
gittalk.

md would produce HTML just different enough to cause the hacked up 'delay'
stuff (using @@ and a bug in slidy) to require an extra click/spacebar-press
at those places; irritating enough not to use it for this!