Doc: replace SHA keyword with digest
The output of Cryptographic hash function is usually called as digest. In
some part of git manual, digest is interchanged with SHA, but SHA is a
cryptographic hash function and git uses SHA-1, which is a part of SHA
algorithm family.

Pic: https://commons.wikimedia.org/wiki/File:Cryptographic_Hash_Function.svg